+++ tests/e2e/e2e_test.go
@@ -466,3 +466,9 @@ func TestE2ERedirectLocationHeader(t *testing.T) {
 	checkRequest(t, "redirect.example.com", envoyEndpoint+"/test-redirect", http.MethodGet, http.StatusOK, false, "", "Referer", "http://172.17.0.1:32855/test-redirect")
 	checkInLogs(t, http.StatusOK, http.MethodGet, "/anything")
 }
+
+func TestE2ERedirectLocationHeaderExpanded(t *testing.T) {
+	backendLogs.Reset()
+	checkRequest(t, "redirect.example.com", envoyEndpoint+"/test-expanded-redirect", http.MethodGet, http.StatusOK, false, "", "Referer", "http://172.17.0.1:32855/test-expanded-redirect")
+	checkInLogs(t, http.StatusOK, http.MethodGet, "/anything\\?uri=/test-expanded-redirect")
+}